Add files via upload
Array and List comparison done

+++ SQ.h
@@ -27,7 +27,7 @@ namespace StationaryQueue
     {
     private:
     public:
-        short int offset;
+        int offset;
         deque<T> *data;
         void offsetmove(bool back = false)
         {
